refactoring and first pass on model generation
