+++ BondGraphTools/components/base.json
@@ -4,6 +4,7 @@
   "components": {
     "0": {
       "name": "Equipotential Node",
+      "class": "BaseComponent",
       "ports": {
         "i": null
       },
@@ -14,6 +15,7 @@
     },
     "1": {
       "name": "Equal Flow Node",
+      "class": "BaseComponent",
       "ports": {
         "i":null
       },
@@ -24,6 +26,7 @@
     },
     "R": {
       "name": "Generalised Linear Resistor",
+      "class": "BaseComponent",
       "ports": {
         "0": null
       },
@@ -43,6 +46,7 @@
     },
     "C": {
       "name": "Generalised Linear Capacitor",
+      "class": "BaseComponent",
       "ports": {
         "0": null
       },
@@ -65,6 +69,7 @@
     },
     "I": {
       "name": "Generalised Linear Inductor",
+      "class": "BaseComponent",
       "ports": {
         "0": null
       },
@@ -87,6 +92,7 @@
     },
     "Se": {
       "name": "Effort Source",
+      "class": "BaseComponent",
       "ports": {
         "0": null
       },
@@ -99,6 +105,7 @@
     },
     "Sf": {
       "name": "Flow Source",
+      "class": "BaseComponent",
       "ports": {
         "0": null
       },
@@ -111,6 +118,7 @@
     },
     "TF": {
       "name": "Linear Transformer",
+      "class": "BaseComponent",
       "ports": {
         "0": {
           "name": "Primary"
@@ -132,6 +140,7 @@
     },
     "GY": {
       "name": "Linear Gyrator",
+      "class": "BaseComponent",
       "ports": {
         "0": {
           "name": "Primary"
