+++ BondGraphTools/compound.py
@@ -1,162 +1,7 @@
-"""
-Bond Graph Model base files.
-"""
 
-import logging
-import sympy
-import copy
-
-from .view import GraphLayout, Glyph
-from .component_manager import get_component, base_id
-
-logger = logging.getLogger(__name__)
-
-
-def new(component, name=None, library=base_id, value=None):
-    """
-    Creates a new Bond Graph from a library component.
-
-    Args:
-        component(str or obj): The type of component to create.
-         If a string is specified, the the component will be created from the
-         appropriate libaray. If an existing bond graph is given, the bond
-         graph will be cloned.
-        name (str): The name for the new component
-        library (str): The library from which to find this component (if
-        component is specified by string).
-
-    Returns: instance of :obj:`BondGraph`
-
-    """
-    if isinstance(component, str):
-        build_args = get_component(component, library)
-
-        if name:
-            build_args.update({"name": name})
-        if value:
-            _update_build_params(build_args, value)
-
-        return AtomicComponent(type=component, **build_args)
-
-    elif isinstance(component, BondGraphBase):
-        obj = copy.copy(component)
-        if name:
-            obj.name = name
-        if value:
-            _update_build_params(obj.__dict__, value)
-
-        return obj
-
-    else:
-        raise NotImplementedError(
-            "New not implemented for object {}", component
-        )
-
-
-def _update_build_params(build_args, value):
-    if isinstance(value, (list, tuple)):
-        assignments = zip(value, build_args["params"].keys())
-        for param, v in assignments:
-            build_args["params"][param] = v
-    elif isinstance(value, dict):
-        for param, v in value:
-            build_args["params"][param] = v
-    else:
-        p = next(iter(build_args["params"]))
-        build_args["params"][p] = value
-
-
-class BondGraphBase:
-    def __init__(self, name, parent=None,
-                 ports=None, params=None):
-        """
-        Base class definition for all bond graphs.
-
-        Args:
-            name: Assumed to be unique
-            metadata (dict):
-        """
-        #super().__init__(name, parent)
-        self.name = name
-        self._ports = ports
-        """ List of exposed Power ports"""
-
-        """ Dictionary of internal parameter and their values. The key is 
-        the internal parameter, the value may be an exposed control value,
-        a function of time, or a constant."""
-        self.view = None
-
-    @property
-    def ports(self):
-        return self._ports
-
-    @property
-    def state_vars(self):
-        return NotImplementedError
-
-    @property
-    def control_vars(self):
-        return NotImplementedError
-
-    @property
-    def params(self):
-        raise NotImplementedError
-
-    def __hash__(self):
-        return id(self)
-
-    def __eq__(self, other):
-        return self.__dict__ == other.__dict__
-
-
-class AtomicComponent(BondGraphBase):
-    """
-    Atomic bond graph components are those defined by constitutive relations.
-    """
-    def __init__(self, type, constitutive_relations,
-                 state_vars=None, params=None, **kwargs):
-        super().__init__(**kwargs)
-
-        self._state_vars = state_vars
-
-        self._params = params
-
-        self.view = Glyph()
-        self.type = type
-        self.constitutive_relations = constitutive_relations
-
-    def __eq__(self, other):
-        return self.__dict__ == self.__dict__
-
-    @property
-    def control_vars(self):
-        if self.params:
-            return [param for param, value in self.params.items() if not value]
-        else:
-            return []
-
-    @property
-    def params(self):
-        return self._params
-
-    @property
-    def state_vars(self):
-        return self._state_vars
-
-    @property
-    def model(self):
-        model = []
-        for rel_str in self.constitutive_relations:
-            model.append(sympy.sympify(rel_str))
-
-    def __add__(self, other):
-        return BondGraph(
-            name="{}+{}".format(self.name, other.name),
-            components=[self, other]
-        )
-
-    def __hash__(self):
-        return id(self)
+from .base import BondGraphBase, InvalidPortException, \
+    InvalidComponentException
+from .view import GraphLayout
 
 
 class BondGraph(BondGraphBase):
@@ -347,14 +192,5 @@ class BondGraph(BondGraphBase):
                        or (cmp((c1, port_1), dest) and cmp((c2, port_2), src))
 
         self.bonds = [bond for bond in self.bonds if not is_target(*bond)]
-
-
-class InvalidPortException(Exception):
-    pass
-
-
-class InvalidComponentException(Exception):
-    pass
-
